moved util functions from global to where they make a bit more sense
