+++ src/item.h
@@ -9,7 +9,7 @@
 #ifndef ITEM_H
 #define ITEM_H
 
-#include "tscomplex.h"
+#include "util.h"
 
 typedef struct Item Item;
