--- src/tscomplex.h
@@ -1,22 +1,0 @@
-/*
- * This software is licensed under the terms of the MIT-License
- * See COPYING for further information.
- * ---
- * Copyright (C) 2011, Lukas Weber <laochailan@web.de>
- */
-
-#ifndef TSCOMPLEX_H
-#define TSCOMPLEX_H
-
-#include <complex.h>
-
-// This is a workaround to properly specify the type of our "complex" variables...
-// Taisei code always uses just "complex" when it actually means "complex double", which is not really correct...
-// gcc doesn't seem to care, other compilers do (e.g. clang)
-
-#ifdef complex
-    #undef complex
-    #define complex _Complex double
-#endif
-
-#endif
